requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
openai==1.3.0
transformers==4.35.0
# torch==2.1.0 # Commented out due to installation issues
scikit-learn==1.3.0
numpy==1.24.0
pandas==2.0.3
cryptography==41.0.0
pyjwt==2.8.0
requests==2.31.0
boto3==1.28.0
prometheus-client==0.17.1
elasticsearch==8.10.0
black==23.9.0
flake8==6.1.0
